create InUseImageProvider
