+++ dbt_platform_helper/commands/ecr_housekeeping.py
@@ -2,9 +2,9 @@ import click
 
 from dbt_platform_helper.domain.ecr_housekeeping import ECRHousekeeping
 from dbt_platform_helper.domain.ecr_housekeeping import ImageProvider
-from dbt_platform_helper.domain.ecr_housekeeping import InUseImageProvider
 from dbt_platform_helper.domain.versioning import PlatformHelperVersioning
 from dbt_platform_helper.platform_exception import PlatformException
+from dbt_platform_helper.providers.in_use_image_provider import InUseImageProvider
 from dbt_platform_helper.providers.io import ClickIOProvider
 from dbt_platform_helper.utils.aws import get_aws_session_or_abort
 from dbt_platform_helper.utils.click import ClickDocOptGroup
